// src/types/ClientData.ts
export type ClientData = {
  nomeAutor?: string;
  dataNascimento?: string;
  estadoCivil?: string;
  profissao?: string;
  cpf?: string;
  rg?: string;
  enderecoCompleto?: string;
  cep?: string;
  emailAutor?: string;
  enderecoEscritorio?: string;
  enderecoDfIprev?: string;
  valorDanoMoral?: string;
  valorDaCausa?: string;
  nomeAdvogado?: string;
  ufOab?: string;
  numeroOab?: string;
  cidadeUf?: string;
};